+++ src/datapoints-graph/charts/echarts-options.service.ts
@@ -6,6 +6,7 @@ import {
   DatapointChartRenderType,
   DatapointWithValues,
   DateString,
+  SeriesDatapointInfo,
   SeriesValue,
 } from '../model';
 import { YAxisService } from './y-axis.service';
@@ -77,6 +78,11 @@ export class EchartsOptionsService {
       },
       legend: {
         show: false,
+        // legend styling is needed for screenshot feature which adds legend to image
+        itemHeight: 8,
+        textStyle: {
+          fontSize: 10,
+        },
       },
       xAxis: {
         min: timeRange.dateFrom,
@@ -130,13 +136,16 @@ export class EchartsOptionsService {
     renderType: Exclude<DatapointChartRenderType, 'area'>,
     idx: number,
     isMinMaxChart = false
-  ): SeriesOption & { datapointId: string } {
+  ): SeriesOption & SeriesDatapointInfo {
     const datapointId = dp.__target.id + dp.fragment + dp.series;
     return {
       datapointId,
+      datapointUnit: dp.unit,
       // 'id' property is needed as 'seriesId' in tooltip formatter
       id: isMinMaxChart ? `${datapointId}/${renderType}` : `${datapointId}`,
-      name: dp.label,
+      name: `${dp.label} (${dp.__target.name})`,
+      // datapointLabel used to proper display of tooltip
+      datapointLabel: dp.label,
       data: Object.entries(dp.values).map(([dateString, values]) => {
         return [dateString, values[0][renderType]];
       }),
@@ -170,6 +179,7 @@ export class EchartsOptionsService {
           );
           value =
             `${minValue[1]} — ${maxValue[1]}` +
+            (series.datapointUnit ? ` ${series.datapointUnit}` : '') +
             `<div style="font-size: 11px">${this.datePipe.transform(
               minValue[0]
             )}</div>`;
@@ -186,6 +196,7 @@ export class EchartsOptionsService {
           }
           value =
             seriesValue[1]?.toString() +
+            (series.datapointUnit ? ` ${series.datapointUnit}` : '') +
             `<div style="font-size: 11px">${this.datePipe.transform(
               seriesValue[0]
             )}</div>`;
@@ -193,7 +204,7 @@ export class EchartsOptionsService {
 
         YAxisReadings.push(
           `<span style='display: inline-block; background-color: ${series.itemStyle.color} ; height: 12px; width: 12px; border-radius: 50%; margin-right: 4px;'></span>` + // color circle
-            `<strong>${series.name}: </strong>` + // name
+            `<strong>${series.datapointLabel}: </strong>` + // name
             value // single value or min-max range
         );
       });